Segunda clase/gui_rodrigo/requirements.txt
pillow
pygame